Tests/LibWeb/Text/input/css/function-parameter-default-value-asf.html

<!doctype html>
<style>
    @function --var-default(--x: var(--fallback)) {
    }

    @function --invalid-var-default(--x: var(, 1px)) {
    }
</style>
<script src="../include.js"></script>
<script>
    test(() => {
        for (const rule of document.styleSheets[0].cssRules) {
            println(rule.cssText);
        }
    });
</script>
