Tests/LibWeb/Text/expected/css/function-parameter-default-value-asf.txt

@function --var-default(--x: var(--fallback)) { }
